Snakes Partially Work (Unstable!)
Snakes kind of work!  The implementation is not ideal, and I'll refactor it later when I have a better large scale understanding of the system.  There is one snake-related mechanic which doesn't work yet - when you push on a snake from two places at once, at least 4 blocks apart, the snake should "break", and currently this causes the game to crash.