Fixed Snakes (maybe for good, this time)
I think I finally got the last bug out of the movement algorithm - at the very least, I haven't run into any unexpected behavior in the current build.  That said, the algorithm still needs to be structured more neatly than it is, because at the moment some parts feel a bit too ad hoc, but the important thing for now is that it (hopefully) works.  If that's correct, then this is the first truly "stable" build in a while.