Reworking Movement System (Stable, not Fully Functional)
The movement algorithm is now done in a somewhat smarter way, with a bit more abstraction to move block type specific details elsewhere.  The WorldMap class has significantly less responsibility now.  The current version of the game won't crash, but that's not because the algorithm is done - currently the link-forming step is skipped, so all of the "mechanics" are trivialized.