Movement Refactoring Finished
Everything should be stable again, but now with a better designed movement algorithm.  In particular, it now definitely is linear in the number of blocks+links, and "global" information is never used except in room init/loading (this will be important when rooms are bigger!).  It's likely that the algorithm can be improved - some of the link checks are definitely overkill - but the important thing is that it seems to be Correct.  Cool.