Creation and Deletion work in Editor
Now it's more like an editor!  In fact, this editor now has every useful feature that the old "editor" had, and technically more (because you have the option to overwrite files).  It's not perfect, but kinks can be worked out later - the next big priority is making camera contexts via the editor