Cleaned up (de)serializing with reflection
Now we don't need quite so much boilerplate to create a new object.  In particular, every new object needs to be included in the ObjCode enum, given serialize and deserialize methods, and incorporated in the GameObject maps (at the bottom of main.cpp).  I'll try to similarly normalize other serialization processes, if it seems helpful.