Fixed up editor input
It's now better, internally and externally.  On the one hand, the game correctly ignores inputs which *should* be eaten by the editor window (rather than creating/deleting things behind it!).  On the other hand, the internal representation of the "tabs" of the editor is much neater than it was before.