Refactoring Done
The editor is now as good as it ever was, but internally it's structured much better than it was before.  In particular, the class RoomManager is finally gone.  The design of the editor GUI is still minimal, and not pretty, but the Door tab is a good example of how temporary game states can be useful.