Maps can be shifted/extended
In other words, their dimensions can be changed after creation, on any of the 6 sides.  Note that this introduces a new liability: any objects which keep track of positions in strange ways need to be able to keep these things in sync.  In particular, Gates need to update the position of retracted GateBodies (and more cases will probably appear later).  Doors are another case which is sensitive to shifts, and a workaround is already in place, but it's not currently being used.